Moved Fighter and Wizard into 'players' package
Subclasses of Monster are in the monsters package. Moved subclasses of
Player into players package for consistency.
