+++ src/com/dd/entities/players/Wizard.java
@@ -1,7 +1,8 @@
-package com.dd.entities;
+package com.dd.entities.players;
 
 import com.dd.Stats;
 import com.dd.dataTypes.enums.Equip;
+import com.dd.entities.Player;
 import com.dd.exceptions.*;
 import com.dd.items.*;
 import com.dd.levels.MapPosition;
